Refactor /id api and commands within the cluster api
