--- backend/apid/actions/cluster_id_test.go
@@ -1,82 +1,0 @@
-package actions
-
-import (
-	"context"
-	"errors"
-	"testing"
-
-	"github.com/google/uuid"
-	"github.com/sensu/sensu-go/backend/store"
-	"github.com/sensu/sensu-go/testing/mockstore"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-)
-
-func TestNewClusterIDController(t *testing.T) {
-	assert := assert.New(t)
-
-	store := &mockstore.MockStore{}
-	actions := NewClusterIDController(store)
-
-	assert.NotNil(actions)
-	assert.Equal(store, actions.store)
-}
-
-func TestGetClusterID(t *testing.T) {
-	testCases := []struct {
-		name            string
-		ctx             context.Context
-		storeErr        error
-		expectedResult  string
-		expectedErr     bool
-		expectedErrCode ErrCode
-	}{
-		{
-			name:           "Get",
-			ctx:            context.Background(),
-			expectedResult: uuid.New().String(),
-		},
-		{
-			name:            "Not found",
-			ctx:             context.Background(),
-			storeErr:        &store.ErrNotFound{},
-			expectedErr:     true,
-			expectedErrCode: NotFound,
-		},
-		{
-			name:            "Store error",
-			ctx:             context.Background(),
-			storeErr:        errors.New("some error"),
-			expectedErr:     true,
-			expectedErrCode: InternalErr,
-		},
-	}
-
-	for _, tc := range testCases {
-		store := &mockstore.MockStore{}
-		actions := NewClusterIDController(store)
-
-		t.Run(tc.name, func(t *testing.T) {
-			assert := assert.New(t)
-
-			store.
-				On("GetClusterID", mock.Anything, mock.Anything).
-				Return(tc.expectedResult, tc.storeErr)
-
-			result, err := actions.Get(tc.ctx)
-
-			if tc.expectedErr {
-				inferErr, ok := err.(Error)
-				if ok {
-					assert.Equal(tc.expectedErrCode, inferErr.Code)
-				} else {
-					assert.Error(err)
-					assert.FailNow("Return value was not of type 'Error'")
-				}
-			} else {
-				assert.NoError(err)
-				assert.Equal(tc.expectedResult, result)
-			}
-		})
-	}
-}
