+++ backend/apid/actions/cluster_test.go
@@ -1,9 +1,15 @@
 package actions
 
 import (
+	"errors"
 	"testing"
 
 	"github.com/coreos/etcd/clientv3"
+	"github.com/google/uuid"
+	"github.com/sensu/sensu-go/backend/store"
+	"github.com/sensu/sensu-go/testing/mockstore"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
 	"golang.org/x/net/context"
 )
 
@@ -29,7 +35,7 @@ func (mockCluster) MemberUpdate(context.Context, uint64, []string) (*clientv3.Me
 var _ clientv3.Cluster = mockCluster{}
 
 func TestMemberList(t *testing.T) {
-	ctrl := NewClusterController(mockCluster{})
+	ctrl := NewClusterController(mockCluster{}, &mockstore.MockStore{})
 
 	_, err := ctrl.MemberList(context.Background())
 	if err != nil {
@@ -38,7 +44,7 @@ func TestMemberList(t *testing.T) {
 }
 
 func TestMemberAdd(t *testing.T) {
-	ctrl := NewClusterController(mockCluster{})
+	ctrl := NewClusterController(mockCluster{}, &mockstore.MockStore{})
 
 	_, err := ctrl.MemberAdd(context.Background(), []string{"foo"})
 	if err != nil {
@@ -47,7 +53,7 @@ func TestMemberAdd(t *testing.T) {
 }
 
 func TestMemberUpdate(t *testing.T) {
-	ctrl := NewClusterController(mockCluster{})
+	ctrl := NewClusterController(mockCluster{}, &mockstore.MockStore{})
 
 	_, err := ctrl.MemberUpdate(context.Background(), 1234, []string{"foo"})
 	if err != nil {
@@ -56,10 +62,79 @@ func TestMemberUpdate(t *testing.T) {
 }
 
 func TestMemberRemove(t *testing.T) {
-	ctrl := NewClusterController(mockCluster{})
+	ctrl := NewClusterController(mockCluster{}, &mockstore.MockStore{})
 
 	_, err := ctrl.MemberRemove(context.Background(), 1234)
 	if err != nil {
 		t.Fatal(err)
+	}
+}
+
+func TestNewClusterIDController(t *testing.T) {
+	assert := assert.New(t)
+
+	store := &mockstore.MockStore{}
+	actions := NewClusterController(mockCluster{}, store)
+
+	assert.NotNil(actions)
+	assert.Equal(store, actions.store)
+}
+
+func TestGetClusterID(t *testing.T) {
+	testCases := []struct {
+		name            string
+		ctx             context.Context
+		storeErr        error
+		expectedResult  string
+		expectedErr     bool
+		expectedErrCode ErrCode
+	}{
+		{
+			name:           "Get",
+			ctx:            context.Background(),
+			expectedResult: uuid.New().String(),
+		},
+		{
+			name:            "Not found",
+			ctx:             context.Background(),
+			storeErr:        &store.ErrNotFound{},
+			expectedErr:     true,
+			expectedErrCode: NotFound,
+		},
+		{
+			name:            "Store error",
+			ctx:             context.Background(),
+			storeErr:        errors.New("some error"),
+			expectedErr:     true,
+			expectedErrCode: InternalErr,
+		},
+	}
+
+	for _, tc := range testCases {
+		store := &mockstore.MockStore{}
+		actions := NewClusterController(mockCluster{}, store)
+
+		t.Run(tc.name, func(t *testing.T) {
+			assert := assert.New(t)
+
+			store.
+				On("GetClusterID", mock.Anything, mock.Anything).
+				Return(tc.expectedResult, tc.storeErr)
+
+			result, err := actions.ClusterID(tc.ctx)
+
+			if tc.expectedErr {
+				inferErr, ok := err.(Error)
+				if ok {
+					assert.Equal(tc.expectedErrCode, inferErr.Code)
+				} else {
+					assert.Error(err)
+					assert.FailNow("Return value was not of type 'Error'")
+				}
+			} else {
+				assert.NoError(err)
+				assert.Equal(tc.expectedResult, result)
+			}
+		})
 	}
 }
