+++ backend/apid/apid.go
@@ -203,7 +203,6 @@ func registerUnauthenticatedResources(
 		routers.NewHealthRouter(actions.NewHealthController(store, cluster, etcdClientTLSConfig)),
 		routers.NewVersionRouter(actions.NewVersionController(clusterVersion)),
 		routers.NewTessenMetricRouter(actions.NewTessenMetricController(bus)),
-		routers.NewClusterIDRouter(actions.NewClusterIDController(store)),
 	)
 }
 
@@ -260,7 +259,7 @@ func (a *APId) registerRestrictedResources(router *mux.Router) {
 		routers.NewChecksRouter(a.store, a.queueGetter),
 		routers.NewClusterRolesRouter(a.store),
 		routers.NewClusterRoleBindingsRouter(a.store),
-		routers.NewClusterRouter(actions.NewClusterController(a.cluster)),
+		routers.NewClusterRouter(actions.NewClusterController(a.cluster, a.store)),
 		routers.NewEntitiesRouter(a.store),
 		routers.NewEventFiltersRouter(a.store),
 		routers.NewEventsRouter(a.eventStore, a.bus),
