+++ cli/client/cluster.go
@@ -10,7 +10,9 @@ import (
 )
 
 var clusterMembersPath = CreateBasePath(coreAPIGroup, coreAPIVersion, "cluster", "members")
+var clusterIDPath = CreateBasePath(coreAPIGroup, coreAPIVersion, "cluster", "id")
 
+// MemberList lists all members in the cluster.
 func (c *RestClient) MemberList() (*clientv3.MemberListResponse, error) {
 	path := clusterMembersPath()
 	res, err := c.R().Get(path)
@@ -24,6 +26,7 @@ func (c *RestClient) MemberList() (*clientv3.MemberListResponse, error) {
 	return &result, json.Unmarshal(res.Body(), &result)
 }
 
+// MemberAdd adds a member to the cluster.
 func (c *RestClient) MemberAdd(peerAddrs []string) (*clientv3.MemberAddResponse, error) {
 	values := url.Values{"peer-addrs": {strings.Join(peerAddrs, ",")}}.Encode()
 	endpoint := fmt.Sprintf("%s?%s", clusterMembersPath(), values)
@@ -38,6 +41,7 @@ func (c *RestClient) MemberAdd(peerAddrs []string) (*clientv3.MemberAddResponse,
 	return &result, json.Unmarshal(res.Body(), &result)
 }
 
+// MemberUpdate updates a member in the cluster.
 func (c *RestClient) MemberUpdate(id uint64, peerAddrs []string) (*clientv3.MemberUpdateResponse, error) {
 	values := url.Values{"peer-addrs": {strings.Join(peerAddrs, ",")}}.Encode()
 	endpoint := fmt.Sprintf("%s/%x?%s", clusterMembersPath(), id, values)
@@ -52,6 +56,7 @@ func (c *RestClient) MemberUpdate(id uint64, peerAddrs []string) (*clientv3.Memb
 	return &result, json.Unmarshal(res.Body(), &result)
 }
 
+// MemberRemove removes a member from the cluster.
 func (c *RestClient) MemberRemove(id uint64) (*clientv3.MemberRemoveResponse, error) {
 	endpoint := fmt.Sprintf("%s/%x", clusterMembersPath(), id)
 	res, err := c.R().Delete(endpoint)
@@ -63,4 +68,18 @@ func (c *RestClient) MemberRemove(id uint64) (*clientv3.MemberRemoveResponse, er
 	}
 	var result clientv3.MemberRemoveResponse
 	return &result, json.Unmarshal(res.Body(), &result)
+}
+
+// FetchClusterID fetches the sensu cluster id.
+func (c *RestClient) FetchClusterID() (string, error) {
+	res, err := c.R().Get(clusterIDPath())
+	if err != nil {
+		return "", fmt.Errorf("GET %q: %s", clusterIDPath(), err)
+	}
+
+	if res.StatusCode() >= 400 {
+		return "", UnmarshalError(res)
+	}
+
+	return string(res.Body()), err
 }
