+++ cli/client/interface.go
@@ -27,7 +27,6 @@ type APIClient interface {
 	AuthenticationAPIClient
 	AssetAPIClient
 	CheckAPIClient
-	ClusterIDAPIClient
 	ClusterRoleAPIClient
 	ClusterRoleBindingAPIClient
 	EntityAPIClient
@@ -92,11 +91,6 @@ type CheckAPIClient interface {
 
 	AddCheckHook(check *types.CheckConfig, checkHook *types.HookList) error
 	RemoveCheckHook(check *types.CheckConfig, checkHookType string, hookName string) error
-}
-
-// ClusterIDAPIClient client methods for sensu cluster id
-type ClusterIDAPIClient interface {
-	FetchClusterID() (string, error)
 }
 
 // ClusterRoleAPIClient client methods for cluster roles
@@ -241,19 +235,22 @@ type SilencedAPIClient interface {
 	UpdateSilenced(*types.Silenced) error
 }
 
-// ClusterMemberClient specifies client methods for cluster membership management
+// ClusterMemberClient specifies client methods for cluster membership management.
 type ClusterMemberClient interface {
-	// MemberList lists cluster members
+	// MemberList lists cluster members.
 	MemberList() (*clientv3.MemberListResponse, error)
 
-	// MemberAdd adds a cluster member
+	// MemberAdd adds a cluster member.
 	MemberAdd(peerAddrs []string) (*clientv3.MemberAddResponse, error)
 
-	// MemberUpdate updates a cluster member
+	// MemberUpdate updates a cluster member.
 	MemberUpdate(id uint64, peerAddrs []string) (*clientv3.MemberUpdateResponse, error)
 
-	// MemberRemove removes a cluster member
+	// MemberRemove removes a cluster member.
 	MemberRemove(id uint64) (*clientv3.MemberRemoveResponse, error)
+
+	// FetchClusterID gets the sensu cluster id.
+	FetchClusterID() (string, error)
 }
 
 // LicenseClient specifies the enteprise client methods for license management.
