--- cli/client/testing/mock_cluster_id.go
@@ -1,7 +1,0 @@
-package testing
-
-// FetchClusterID ...
-func (c *MockClient) FetchClusterID() (string, error) {
-	args := c.Called()
-	return args.Get(0).(string), args.Error(1)
-}
