+++ cli/commands/cluster/member-list.go
@@ -24,7 +24,7 @@ func MemberListCommand(cli *cli.SensuCli) *cobra.Command {
 			if err != nil {
 				return fmt.Errorf("error listing cluster members: %s", err)
 			}
-			err = helpers.PrintTitle(helpers.GetChangedStringValueFlag("format", cmd.Flags()), cli.Config.Format(), fmt.Sprintf("Cluster ID: %x", result.Header.ClusterId), cmd.OutOrStdout())
+			err = helpers.PrintTitle(helpers.GetChangedStringValueFlag("format", cmd.Flags()), cli.Config.Format(), fmt.Sprintf("Etcd Cluster ID: %x", result.Header.ClusterId), cmd.OutOrStdout())
 			if err != nil {
 				return err
 			}
