+++ cli/commands/commands.go
@@ -17,7 +17,6 @@ import (
 	"github.com/sensu/sensu-go/cli/commands/filter"
 	"github.com/sensu/sensu-go/cli/commands/handler"
 	"github.com/sensu/sensu-go/cli/commands/hook"
-	"github.com/sensu/sensu-go/cli/commands/id"
 	"github.com/sensu/sensu-go/cli/commands/logout"
 	"github.com/sensu/sensu-go/cli/commands/mutator"
 	"github.com/sensu/sensu-go/cli/commands/namespace"
@@ -35,7 +34,6 @@ func AddCommands(rootCmd *cobra.Command, cli *cli.SensuCli) {
 		configure.Command(cli),
 		completion.Command(rootCmd),
 		logout.Command(cli),
-		id.Command(cli),
 
 		// Management Commands
 		asset.HelpCommand(cli),
